server/ml_service/requirements.txt
yfinance
pandas
numpy
xgboost
scikit-learn
pymongo
python-dotenv
ta
